split sync function

+++ foldersync/sync/sync.cs
@@ -69,6 +69,7 @@ namespace foldersync.sync
                 switch (currentFileRelative.Value)
                 {
                     case tracker.changeType.removed:
+                        if (!File.Exists(currentFileDestinationAbsolute)) break;
                         File.Delete(currentFileDestinationAbsolute);
                         break;
                     case tracker.changeType.added:
@@ -78,6 +79,45 @@ namespace foldersync.sync
                 }
             }
 
+        }
+        private void syncFolders()
+        {
+            var folderChanges = folderTracker.GetChanges();
+
+            foreach (var currentFolderRelative in folderChanges)
+            {
+                var currentFolderAbsolte = Path.GetFullPath(currentFolderRelative.Key, destinationFolder);
+                if (currentFolderRelative.Value == tracker.changeType.removed && Directory.Exists(currentFolderAbsolte))
+                {
+                    Directory.Delete(currentFolderAbsolte, true);
+                }
+                else if (currentFolderRelative.Value == tracker.changeType.added)
+                {
+                    Directory.CreateDirectory(currentFolderAbsolte);
+                }
+            }
+        }
+
+        private void syncFiles()
+        {
+            var fileChanges = fileTracker.GetChanges();
+
+            foreach (var currentFileRelative in fileChanges)
+            {
+                var currentFileDestinationAbsolute = Path.GetFullPath(currentFileRelative.Key, destinationFolder);
+                var currentFileSourceAbsolute = Path.GetFullPath(currentFileRelative.Key, sourceFolder);
+                switch (currentFileRelative.Value)
+                {
+                    case tracker.changeType.removed:
+                        if (!File.Exists(currentFileDestinationAbsolute)) break;
+                        File.Delete(currentFileDestinationAbsolute);
+                        break;
+                    case tracker.changeType.added:
+                    case tracker.changeType.modified:
+                        File.Copy(currentFileSourceAbsolute, currentFileDestinationAbsolute, true);
+                        break;
+                }
+            }
         }
 
     }
